+++ lib/src/internal/chat_method_keys.dart
@@ -24,6 +24,16 @@ class ChatMethodKeys {
   static const String onMultiDeviceEvent = "onMultiDeviceEvent";
   static const String onConnected = "onConnected";
   static const String onDisconnected = "onDisconnected";
+  static const String onUserDidLoginFromOtherDevice =
+      "onUserDidLoginFromOtherDevice";
+  static const String onUserDidRemoveFromServer = "onUserDidRemoveFromServer";
+  static const String onUserDidForbidByServer = "onUserDidForbidByServer";
+  static const String onUserDidChangePassword = "onUserDidChangePassword";
+  static const String onUserDidLoginTooManyDevice =
+      "onUserDidLoginTooManyDevice";
+  static const String onUserKickedByOtherDevice = "onUserKickedByOtherDevice";
+  static const String onUserAuthenticationFailed = "onUserAuthenticationFailed";
+
   static const String onSendDataToFlutter = "onSendDataToFlutter";
   static const String onTokenWillExpire = "onTokenWillExpire";
   static const String onTokenDidExpire = "onTokenDidExpire";